faq role_admin

+++ src/Controller/FaqController.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
+
 #[Route('/faq')]
 final class FaqController extends AbstractController
 {
@@ -25,6 +26,9 @@ final class FaqController extends AbstractController
     #[Route('/new', name: 'app_faq_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $faq = new Faq();
         $form = $this->createForm(FaqType::class, $faq);
         $form->handleRequest($request);
@@ -53,6 +57,9 @@ final class FaqController extends AbstractController
     #[Route('/{id}/edit', name: 'app_faq_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Faq $faq, EntityManagerInterface $entityManager): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(FaqType::class, $faq);
         $form->handleRequest($request);
 
@@ -71,7 +78,11 @@ final class FaqController extends AbstractController
     #[Route('/{id}', name: 'app_faq_delete', methods: ['POST'])]
     public function delete(Request $request, Faq $faq, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$faq->getId(), $request->getPayload()->getString('_token'))) {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+
+        if ($this->isCsrfTokenValid('delete' . $faq->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($faq);
             $entityManager->flush();
         }
